pyproject.toml
[project]
name = "tta-bot"
version = "0.1.0"
description = "Telegram bot for handling complaints via FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.3.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
